+++ CHANGELOG.md
@@ -10,6 +10,9 @@
   * Swich to Spotifyd Spotify client
   * Add MPRIS metadata/command interface
 
+* Streams
+  * Add LMS Client
+
 ## 0.1.9
 * Web App
   * Add link for text log